refactor: spice: move tests out of large files. (#14132)
This is a follow up to the suggestion from
https://github.com/near/nearcore/pull/14031#discussion_r2262337884

Since now spice tests run as part of CI there should be no concerns with
moving them to separate files.

While at it I also renamed mod for chunk_executor_actor tests to make
rule in Justfile to run spice tests simpler.
