+++ chain/client/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod spice_chunk_executor_actor;
+mod spice_chunk_validator_actor;
